Rename routerAction to routeAction, fix references
